+++ cfluid/base/base_of_client.h
@@ -6,7 +6,7 @@
 
 struct base_of_client {
     struct base_of_handler ofh;
-    int id;
+    uint64_t id;
     int blocking;
     int nconn;
     struct ev_loop *evloop;
@@ -16,7 +16,7 @@ struct base_of_client {
 };
 
 void base_of_client_init(struct base_of_client *oc, 
-                         int id, const char *address, int port);
+                         uint64_t id, const char *address, int port);
 void base_of_client_clean(struct base_of_client *oc);
 int base_of_client_start(struct base_of_client *conn, int block);
 void base_of_client_stop(struct base_of_client *conn);
